requirements.txt
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==2.0.1
itsdangerous==1.1.0
Werkzeug==2.0.3
python-dotenv==1.0.1
Requests==2.32.3